requirements.txt
django
django-cors-headers
djangorestframework
whitenoise
attrs
daphne
channels==3.0.5
pillow
drf-spectacular